pyproject.toml
[project]
name = "inverse-cai"
version = "0.2.0"
authors = [
  { name="rdnfn", email="hi@arduin.io" },
  { name="timokau", email="" },
]
description = "Inverse Constitutional AI (ICAI) pipeline for interpreting pairwise preferences"
readme = "docs/pypi_readme.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "langchain",
  "langchain-community",
  "langchain-openai",
  "langchain-anthropic",
  "anthropic",
  "alpaca-eval<=0.6.5",
  "datasets", # To fetch huggingface datasets
  "scikit-learn", # used for clustering
  "pandas",
  "python-dotenv",
  "loguru",
  "hydra-core",
  "matplotlib",
  "setuptools", # for alpaca-eval
]

[project.optional-dependencies]
dev = [
    "pytest >=6.0",
    "black ==24.10.0",
    "bump-my-version",
]
app = [
    "gradio",
    "plotly",
    "nbconvert",
    "jupyter",
    "ipywidgets", # IProgress in jupyter notebooks
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel.force-include]
"src/inverse_cai/assets" = "inverse_cai/assets"

[project.scripts]
icai-exp = "inverse_cai.experiment.core:run"
icai-app = "inverse_cai.app.main:run"
icai-clean = "inverse_cai.tools.data_cleaner:run"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["src"]
python_files = ["*_test.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "api: marks tests as API tests (deselect with '-m \"not api\"')",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.bumpversion]
current_version = "0.2.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
moveable_tags = []
commit_args = ""
setup_hooks = []
pre_commit_hooks = []
post_commit_hooks = []